9/21 기능수정 1. 웨이브가 끝났을 때는 타워가 굳이 Targeting을 할필요가 없으므로 TowerManager에 List<Tower>로 타워설치할때 추가해놓고 웨이브가 끝났을때 스크립트를 꺼주고 시작할 때 다시 다 켜주게 했다 2. 웨이브 도중에 설치를 할 수도 있기 때문에 처음부터 꺼놓으면 꼬일거 같아서 일단 설치는 켜놓는걸 디폴트로 해놓았다. 어차피 웨이브가 끝나면 다 끄기 때문에

+++ TowerDefense/Assets/Scripts/PoolObjectControl/PoolObjectManager.cs
@@ -13,8 +13,8 @@ namespace PoolObjectControl
         {
             public PoolObjectKey poolObjectKey;
             public GameObject prefab;
-            public int initSize;
-            public int maxSize;
+            public byte initSize;
+            public byte maxSize;
 
             public int CompareTo(Pool other)
             {
